+++ wa/starter/src/main/resources/wa.properties
@@ -21,7 +21,7 @@ useGZIPCompression=true
 
 # Conf directories
 conf.directory=${conf.directory}
-cas.standalone.configurationDirectory=${conf.directory}
+cas.standalone.configuration-directory=${conf.directory}
 cas.authn.oidc.jwks.jwksFile=file:${conf.directory}/oidc.keystore
 
 cas.server.name=http://localhost:8080
